refactor(pipelines,workers): rename base to match respective component
